+++ shared/subgraph.cuh
@@ -11,24 +11,17 @@ class Subgraph
 private:
 
 public:
-	uint num_nodes;
-	uint num_edges;
-	uint numActiveNodes;
-	
-	uint *activeNodes;
-	uint *activeNodesPointer;
-	E *activeEdgeList;
-	
-	uint *d_activeNodes;
-	uint *d_activeNodesPointer;
-	E *d_activeEdgeList;
-	
-	ull max_partition_size;
-	
-	Subgraph(uint num_nodes, uint num_edges);
+    uint num_nodes;
+    ull num_edges;
+    uint numActiveNodes;
+    uint *activeNodes;
+    ull *activeNodesPointer;
+    E *activeEdgeList;
+    uint *d_activeNodes;
+    ull *d_activeNodesPointer;
+    E *d_activeEdgeList;
+    ull max_partition_size;
+    Subgraph(uint num_nodes, ull num_edges);
 };
 
-#endif	//	SUBGRAPH_HPP
-
-
-
+#endif    //    SUBGRAPH_HPP
